selftests/hid: wacom: Test for proper Bluetooth timestamp values
The Bluetooth interface on the CTL-4100 (and some other tablets) sends
batched reports which contain multiple snapshots of the tablet state.
For a long time we forwarded all of these events to userspace without
making any adjustments to the timestamps. This caused problems for
anything that depended on an accurate view of when the events were
actually generated. An early version of the kernel patch which added
support for modifying the timestamps had a bug which caused the
timestamp to be incorrectly calculated in some situations. This commit
adds a test to ensure we don't accidentally break the timestamp behavior
again in the future.

These tests ensure basic sanity of the timestamp stream. Timestamps are
required to monotonically increase with each SYN_REPORT and must not
advance in time by more than a relatively small amount (set to 1 second
for the moment -- much larger than typcial, but small enough to ensure
the test does not fail simply because of a slow / overloaded system).

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>