selftests/hid: wacom: Support batched data in create_report
Some Wacom devices send updates in batches, with a single report housing
one or more snapshots of recent device state. This commit adds logic for
supporting batched data to create_report. If the function is called with
lists of data for x, y, pressure, etc. then we'll feed the parent
function with a list of objects whose attributes have been appropriately
filled in.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>